added condition for deleting contacts from storage

+++ lib/modules/contacts/screens/contacts_list/components/select_app_bar.dart
@@ -1,9 +1,12 @@
 import 'package:aurora_mail/generated/l10n.dart';
+import 'package:aurora_mail/modules/auth/blocs/auth_bloc/auth_bloc.dart';
 import 'package:aurora_mail/modules/contacts/blocs/contacts_bloc/bloc.dart';
 import 'package:aurora_mail/modules/contacts/contacts_domain/models/contact_model.dart';
+import 'package:aurora_mail/modules/contacts/contacts_domain/models/contacts_storage_model.dart';
 import 'package:aurora_mail/modules/contacts/screens/components/groups_select_dialog.dart';
 import 'package:aurora_mail/modules/mail/screens/messages_list/components/selection_controller.dart';
 import 'package:aurora_mail/shared_ui/confirmation_dialog.dart';
+import 'package:collection/collection.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
@@ -18,6 +21,9 @@ class SelectAppBar extends StatelessWidget {
     return BlocBuilder<ContactsBloc, ContactsState>(
       buildWhen: (prev, current) => prev.selectedGroup != current.selectedGroup,
       builder: (context, state) {
+        final areContactsCanBeDeleted = _checkIfContactsCanBeDeleted(
+            state.storages.firstWhereOrNull((e) => e.name == state.selectedStorage),
+            context);
         return AppBar(
           leading: IconButton(
             icon: Icon(
@@ -31,20 +37,30 @@ class SelectAppBar extends StatelessWidget {
               icon: Icon(Icons.drive_file_move_outline),
               onPressed: () => _moveToGroup(context),
             ),
-            if (state.selectedGroup == null)
+            IconButton(
+              icon: Icon(Icons.folder_delete_outlined),
+              onPressed: () => _removeFromGroup(context),
+            ),
+            if (state.selectedGroup == null && areContactsCanBeDeleted)
               IconButton(
                 icon: Icon(Icons.delete_outline),
                 onPressed: () => _deleteContacts(context),
-              )
-            else
-              IconButton(
-                icon: Icon(Icons.folder_delete_outlined),
-                onPressed: () => _removeFromGroup(context),
-              )
+              ),
           ],
         );
       },
     );
+  }
+
+  bool _checkIfContactsCanBeDeleted(
+      ContactsStorage? storage, BuildContext context) {
+    if (storage == null){
+      return false;
+    }
+    final authBlocState = BlocProvider.of<AuthBloc>(context).currentUser;
+    return authBlocState.emailFromLogin == storage.ownerMail
+        ? true
+        : storage.isShared == true && storage.accessCode == 1;
   }
 
   void _deleteContacts(BuildContext context) async {
